Moved resource type item classes
- Moved all resource type item classes to higher up the structure
