+++ app/Validators/Request/Parameters.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Validators\Request;
 
-use App\Item\ItemInterfaceFactory;
+use App\Item\Factory;
 use App\Models\Category;
 use App\Models\ResourceType;
 use App\Models\Subcategory;
@@ -139,13 +139,13 @@ class Parameters
                         $max_year_limit = intval(Date('Y'));
 
                         if ($resource_type_id !== null && $resource_id === null) {
-                            $item_interface = ItemInterfaceFactory::resourceTypeItem($resource_type_id);
+                            $item_interface = Factory::resourceTypeItem($resource_type_id);
                             $min_year_limit = $item_interface->conditionalParameterMinYear($resource_type_id);
                             $max_year_limit = $item_interface->conditionalParameterMaxYear($resource_type_id);
                         }
 
                         if ($resource_type_id !== null && $resource_id !== null) {
-                            $item_interface = ItemInterfaceFactory::item($resource_type_id);
+                            $item_interface = Factory::item($resource_type_id);
                             $min_year_limit = $item_interface->conditionalParameterMinYear($resource_id);
                             $max_year_limit = $item_interface->conditionalParameterMaxYear($resource_id);
                         }
